jOSeph 4.1.2.5-A1
Completly changed Core to have a single changing pane rather than many with varying visibility
