+++ jOSeph_4/core/CoreOptions.java
@@ -1,0 +1,5 @@
+package jOSeph_4.core;
+
+public enum CoreOptions {
+	GENERATOR, QUIZ, CALCULATOR, NEW, SETTINGS
+}
