+++ jOSeph_4/core/CorePaneInterface.java
@@ -1,0 +1,9 @@
+package jOSeph_4.core;
+
+import javafx.scene.layout.Pane;
+
+import java.io.IOException;
+
+public interface CorePaneInterface {
+	public Pane getPane() throws IOException;
+}
