--- jOSeph_4/core/New.java
@@ -1,4 +1,0 @@
-package jOSeph_4.core;
-
-public class New {
-}
