+++ jOSeph_4/resources/controllers/core/QuizMenu_Controller.java
@@ -1,0 +1,12 @@
+package jOSeph_4.resources.controllers.core;
+
+import jOSeph_4.Core;
+
+public class QuizMenu_Controller {
+
+
+	public void onQuizStartPressed(){
+		Core_Controller.closeThread();
+		Core.startQuiz();
+	}
+}
