+++ jOSeph_4/resources/fxml/Core.fxml
@@ -1,19 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.Group?>
 <?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ProgressBar?>
 <?import javafx.scene.control.ScrollPane?>
 <?import javafx.scene.layout.BorderPane?>
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.Pane?>
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.shape.Rectangle?>
-<?import javafx.scene.text.Font?>
 
-<BorderPane id="borderpane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" stylesheets="@../css/core/core.css" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="jOSeph_4.resources.controllers.Core_Controller">
+<BorderPane id="borderpane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" stylesheets="@../css/core/core.css" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="jOSeph_4.resources.controllers.core.Core_Controller">
    <left>
       <HBox BorderPane.alignment="CENTER">
          <children>
@@ -37,62 +32,6 @@
       </HBox>
    </left>
    <center>
-      <Group BorderPane.alignment="CENTER">
-         <children>
-            <Pane fx:id="genPane" prefHeight="250.0" prefWidth="250.0" stylesheets="@../css/core/generator.css">
-               <children>
-                  <VBox alignment="CENTER" prefHeight="250.0" prefWidth="250.0" spacing="20.0">
-                     <children>
-                        <Label text="Generating...">
-                           <font>
-                              <Font name="Arial" size="16.0" />
-                           </font>
-                        </Label>
-                        <Button id="generate" mnemonicParsing="false" onAction="#onGeneratorClick" prefHeight="200.0" prefWidth="200.0" text="Generate">
-                           <font>
-                              <Font name="Arial" size="29.0" />
-                           </font>
-                        </Button>
-                        <ProgressBar fx:id="bar" prefWidth="200.0" progress="0.0" />
-                        <HBox alignment="CENTER" prefHeight="100.0" prefWidth="200.0">
-                           <children>
-                              <Label text="Ideas: ">
-                                 <font>
-                                    <Font name="Arial" size="15.0" />
-                                 </font>
-                              </Label>
-                              <Label fx:id="count" text="0">
-                                 <font>
-                                    <Font name="Arial" size="15.0" />
-                                 </font>
-                              </Label>
-                           </children>
-                        </HBox>
-                     </children>
-                     <padding>
-                        <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
-                     </padding>
-                  </VBox>
-               </children>
-            </Pane>
-            <Pane fx:id="quizPane" stylesheets="@../css/core/quiz.css" visible="false">
-               <children>
-                  <VBox alignment="CENTER">
-                     <children>
-                        <Label alignment="TOP_LEFT" text="Welcome to the quiz">
-                           <VBox.margin>
-                              <Insets bottom="40.0" left="40.0" right="40.0" top="40.0" />
-                           </VBox.margin>
-                        </Label>
-                        <Button mnemonicParsing="false" onAction="#onQuizStartPressed" text="Start" />
-                     </children>
-                  </VBox>
-               </children>
-            </Pane>
-            <Pane fx:id="calculatorPane" prefHeight="200.0" prefWidth="200.0" visible="false" />
-            <Pane fx:id="newPane" prefHeight="200.0" prefWidth="200.0" visible="false" />
-            <Pane fx:id="settingsPane" prefHeight="200.0" prefWidth="200.0" visible="false" />
-         </children>
-      </Group>
+      <Pane fx:id="centerPane" prefHeight="200.0" prefWidth="200.0" BorderPane.alignment="CENTER" />
    </center>
 </BorderPane>
