+++ jOSeph_4/resources/fxml/core/Core.fxml
@@ -1,0 +1,98 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.Group?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.ProgressBar?>
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.Pane?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.shape.Rectangle?>
+<?import javafx.scene.text.Font?>
+
+<BorderPane id="borderpane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" stylesheets="@../../css/core/core.css" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="jOSeph_4.resources.controllers.core.Core_Controller">
+   <left>
+      <HBox BorderPane.alignment="CENTER">
+         <children>
+            <ScrollPane id="scroll" hbarPolicy="NEVER" prefWidth="96.0">
+               <content>
+                  <VBox fx:id="vbox" maxWidth="90.0">
+                     <children>
+                        <Button id="generate_s" maxHeight="70.0" maxWidth="128.0" minHeight="70.0" minWidth="64.0" mnemonicParsing="false" onAction="#onGenMenuClick" styleClass="icon" />
+                        <Button id="quiz_s" maxHeight="70.0" maxWidth="128.0" minHeight="70.0" minWidth="64.0" mnemonicParsing="false" onAction="#onQuizMenuClick" styleClass="icon" />
+                        <Button id="calc_s" maxHeight="70.0" maxWidth="128.0" minHeight="70.0" minWidth="64.0" mnemonicParsing="false" onAction="#onCalcMenuClick" styleClass="icon" />
+                        <Button id="new_s" maxHeight="70.0" maxWidth="128.0" minHeight="70.0" minWidth="64.0" mnemonicParsing="false" onAction="#onNewMenuClick" styleClass="icon" />
+                        <Button id="exit_s" maxHeight="70.0" maxWidth="128.0" minHeight="70.0" minWidth="64.0" mnemonicParsing="false" onAction="#exit" styleClass="icon" />
+                        <Button id="next_s" layoutX="10.0" layoutY="360.0" maxHeight="70.0" maxWidth="128.0" minHeight="70.0" minWidth="64.0" mnemonicParsing="false" styleClass="icon" />
+                        <Button id="settings_s" maxHeight="70.0" maxWidth="128.0" minHeight="70.0" minWidth="64.0" mnemonicParsing="false" onAction="#onSettingsMenuClick" styleClass="icon" />
+                        <Rectangle fx:id="rect" arcHeight="5.0" arcWidth="5.0" fill="#282828" stroke="#282828" strokeType="INSIDE" />
+                     </children>
+                  </VBox>
+               </content>
+            </ScrollPane>
+         </children>
+      </HBox>
+   </left>
+   <center>
+      <Group BorderPane.alignment="CENTER">
+         <children>
+            <Pane fx:id="genPane" prefHeight="250.0" prefWidth="250.0" stylesheets="@../../css/core/generator.css">
+               <children>
+                  <VBox alignment="CENTER" prefHeight="250.0" prefWidth="250.0" spacing="20.0">
+                     <children>
+                        <Label text="Generating...">
+                           <font>
+                              <Font name="Arial" size="16.0" />
+                           </font>
+                        </Label>
+                        <Button id="generate" mnemonicParsing="false" onAction="#onGeneratorClick" prefHeight="200.0" prefWidth="200.0" text="Generate">
+                           <font>
+                              <Font name="Arial" size="29.0" />
+                           </font>
+                        </Button>
+                        <ProgressBar fx:id="bar" prefWidth="200.0" progress="0.0" />
+                        <HBox alignment="CENTER" prefHeight="100.0" prefWidth="200.0">
+                           <children>
+                              <Label text="Ideas: ">
+                                 <font>
+                                    <Font name="Arial" size="15.0" />
+                                 </font>
+                              </Label>
+                              <Label fx:id="count" text="0">
+                                 <font>
+                                    <Font name="Arial" size="15.0" />
+                                 </font>
+                              </Label>
+                           </children>
+                        </HBox>
+                     </children>
+                     <padding>
+                        <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
+                     </padding>
+                  </VBox>
+               </children>
+            </Pane>
+            <Pane fx:id="quizPane" stylesheets="@../../css/core/quiz.css" visible="false">
+               <children>
+                  <VBox alignment="CENTER">
+                     <children>
+                        <Label alignment="TOP_LEFT" text="Welcome to the quiz">
+                           <VBox.margin>
+                              <Insets bottom="40.0" left="40.0" right="40.0" top="40.0" />
+                           </VBox.margin>
+                        </Label>
+                        <Button mnemonicParsing="false" onAction="#onQuizStartPressed" text="Start" />
+                     </children>
+                  </VBox>
+               </children>
+            </Pane>
+            <Pane fx:id="calculatorPane" prefHeight="200.0" prefWidth="200.0" visible="false" />
+            <Pane fx:id="newPane" prefHeight="200.0" prefWidth="200.0" visible="false" />
+            <Pane fx:id="settingsPane" prefHeight="200.0" prefWidth="200.0" visible="false" />
+         </children>
+      </Group>
+   </center>
+</BorderPane>
