+++ jOSeph_4/resources/fxml/core/Generator.fxml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.ProgressBar?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.Pane?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+
+<Pane fx:id="genPane" prefHeight="250.0" prefWidth="250.0" stylesheets="@../../css/core/generator.css" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="jOSeph_4.resources.controllers.core.Generator_Controller">
+   <children>
+      <VBox alignment="CENTER" prefHeight="250.0" prefWidth="250.0" spacing="20.0">
+         <children>
+            <Label text="Generating...">
+               <font>
+                  <Font name="Arial" size="16.0" />
+               </font>
+            </Label>
+            <Button id="generate" mnemonicParsing="false" onAction="#onGeneratorClick" prefHeight="200.0" prefWidth="200.0" text="Generate">
+               <font>
+                  <Font name="Arial" size="29.0" />
+               </font>
+            </Button>
+            <ProgressBar fx:id="bar" prefWidth="200.0" progress="0.0" />
+            <HBox alignment="CENTER" prefHeight="100.0" prefWidth="200.0">
+               <children>
+                  <Label text="Ideas: ">
+                     <font>
+                        <Font name="Arial" size="15.0" />
+                     </font>
+                  </Label>
+                  <Label fx:id="count" text="0">
+                     <font>
+                        <Font name="Arial" size="15.0" />
+                     </font>
+                  </Label>
+               </children>
+            </HBox>
+         </children>
+         <padding>
+            <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
+         </padding>
+      </VBox>
+   </children>
+</Pane>
