+++ jOSeph_4/resources/fxml/core/QuizMenu.fxml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.Pane?>
+<?import javafx.scene.layout.VBox?>
+
+<Pane fx:id="quizPane" stylesheets="@../../css/core/quiz.css" xmlns="http://javafx.com/javafx/8.0.111" xmlns:fx="http://javafx.com/fxml/1" fx:controller="jOSeph_4.resources.controllers.core.QuizMenu_Controller">
+   <children>
+      <VBox alignment="CENTER">
+         <children>
+            <Label alignment="TOP_LEFT" text="Welcome to the quiz">
+               <VBox.margin>
+                  <Insets bottom="40.0" left="40.0" right="40.0" top="40.0" />
+               </VBox.margin>
+            </Label>
+            <Button mnemonicParsing="false" onAction="#onQuizStartPressed" text="Start" />
+         </children>
+      </VBox>
+   </children>
+</Pane>
